tools/requirements.txt
jsonschema==4.23.0
pyyaml==6.0.2
pandas==2.2.2
numpy==1.26.4
